# Optimized SAM parser