This simple tweak reduced runtime from 6.86 s ± 25.2 ms to 4.93 s ± 4.12 ms on a real SAM file.

It can be further optimized using function reference reassignment so that it won't check header (@) every time.